refactor away from barrel file so component tests run
