+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/ModalForm.tsx
@@ -6,7 +6,7 @@ import { FormElementTypes, ElementProperties } from "@lib/types";
 
 import { FormElementWithIndex, LocalizedElementProperties } from "@lib/types/form-builder-types";
 import { Checkbox, Input, TextArea, InfoDetails } from "@formBuilder/components/shared";
-import { useTemplateStore } from "@lib/store";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { AutocompleteDropdown } from "./AutocompleteDropdown";
 
 const ModalLabel = ({ children, ...props }: React.LabelHTMLAttributes<HTMLLabelElement>) => (
