+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/ModalFormRules.tsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import { useTranslation } from "@i18n/client";
 
 import { Button } from "@clientComponents/globals";
-import { useTemplateStore } from "@lib/store";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { ModalProperties } from "@lib/store/useModalRulesStore";
 import { FormElementWithIndex } from "@lib/types/form-builder-types";
 import { ChoiceRule } from "@lib/formContext";
