+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/tests/PanelActions.cy.tsx
@@ -1,14 +1,8 @@
 "use client";
 import React from "react";
-// import { PanelActions } from "../PanelActions";
-const PanelActions = () => <></>;
+import { PanelActions } from "../PanelActions";
 
-/**
- * Does not work because PanelActions imports ElementDialog which
- * imports useElementOptions which imports useFlag which imports
- * a server action.
- */
-describe.skip("<PanelActions />", () => {
+describe("<PanelActions />", () => {
   it("enables move buttons for item that is not first or last", () => {
     cy.viewport(800, 80);
     // see: https://on.cypress.io/mounting-react
