+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/publish/Publish.tsx
@@ -2,7 +2,7 @@
 import React, { useCallback, useState } from "react";
 import { useTranslation } from "@i18n/client";
 import { useRouter } from "next/navigation";
-import { useTemplateStore } from "@lib/store";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { useAllowPublish, useRehydrate } from "@lib/hooks/form-builder";
 import { CancelIcon, CircleCheckIcon, LockIcon } from "@serverComponents/icons";
 import { Button, Alert } from "@clientComponents/globals";
