+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/SetClosingDate.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { useCallback, useState } from "react";
 import { useTranslation } from "@i18n/client";
-import { useTemplateStore } from "@lib/store";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { toast } from "@formBuilder/components/shared/Toast";
 import axios from "axios";
