+++ lib/hooks/form-builder/useHandleAdd.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useCallback } from "react";
 import { FormElementTypes } from "@lib/types";
-import { useTemplateStore } from "@lib/store";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { blockLoader, LoaderType } from "../../utils/form-builder/blockLoader";
 import { allowedTemplates } from "@lib/utils/form-builder";
 import { defaultField, createElement, setDescription } from "@lib/utils/form-builder/itemHelper";
