+++ lib/hooks/index.ts
@@ -2,5 +2,4 @@ export { useRefresh } from "./useRefresh";
 export { useExternalScript } from "./useExternalScript";
 export { useFormTimer } from "./useFormTimer";
 export { useAccessControl, AccessControlProvider } from "./useAccessControl";
-export { useIsFormClosed } from "./useIsFormClosed";
 export { useFormValuesChanged } from "./useValueChanged";
