+++ lib/hooks/useIsFormClosed.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useTemplateStore } from "@lib/store";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { dateHasPast } from "@lib/utils";
 
 export const useIsFormClosed = () => {
