+++ lib/store/index.ts
@@ -1,9 +1,2 @@
-export {
-  useTemplateStore,
-  TemplateStoreProvider,
-  clearTemplateStore,
-  useSubscibeToTemplateStore,
-  TemplateStoreContext,
-} from "./useTemplateStore";
 export { useModalStore } from "./useModalStore";
 export { useModalRulesStore } from "./useModalRulesStore";
